+++ lib/c4fm_soft_decision_fc_impl.h
@@ -32,7 +32,7 @@ namespace gr {
       float d_gain;
 
      public:
-      c4fm_soft_decision_fc_impl(double gain);
+      c4fm_soft_decision_fc_impl(float gain);
       ~c4fm_soft_decision_fc_impl();
 
       // Where all the action really happens
